pyproject.toml
[project]
name = "besx"
version = "0.1.0"
description = "Battery Energy Storage Simulator"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    { name = "BESx Team" },
]
dependencies = [
    "numpy",
    "pandas",
    "matplotlib",
    "scipy",
    "openpyxl",
    "rainflow",
    "ipython",
    "colorlog",
    "streamlit",
    "plotly",
    "pydantic",
    "google-genai",
]

[dependency-groups]
dev = [
    "ruff",
    "mypy",
    "pytest",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "C4", "SIM"]
ignore = []

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true
check_untyped_defs = true
disallow_untyped_decorators = false
warn_unused_ignores = true
warn_redundant_casts = true

[tool.hatch.build.targets.wheel]
packages = ["src/besx"]
